+++ tests/TestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Xgenious\XgApiClient\Tests;
+
+use Orchestra\Testbench\TestCase as Orchestra;
+use Xgenious\XgApiClient\XgApiClientServiceProvider;
+
+class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app)
+    {
+        return [
+            XgApiClientServiceProvider::class,
+        ];
+    }
+}